refactor: unify supabase client
