+++ hooks/use-auth-status.ts
@@ -1,7 +1,7 @@
 "use client"
 
 import { useState, useEffect } from "react"
-import { supabase } from "@/lib/supabase-client"
+import { supabase } from "@/lib/supabase/client"
 
 export function useAuthStatus() {
   const [isAuthenticated, setIsAuthenticated] = useState<boolean | null>(null)
