--- lib/supabase.ts
@@ -1,7 +1,0 @@
-import { createClient } from "@supabase/supabase-js"
-
-// Create a single supabase client for interacting with your database
-export const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL || "",
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || "",
-)
